requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
Django==2.0.1
django-filter==1.1.0
djangorestframework==3.7.7
idna==2.6
kombu==4.1.0
pytz==2017.3
redis==2.10.6
requests==2.18.4
urllib3==1.22
vine==1.1.4
